requirements.txt
fastapi
uvicorn
streamlit
requests
pandas
matplotlib
python-dotenv
jinja2